Sqldelight Movie Detail complete

+++ composeApp/src/commonMain/kotlin/org/example/firstcmpproject/movies/MovieItem.kt
@@ -18,7 +18,7 @@ import org.example.firstcmpproject.movies.data.vos.MovieVO
 import org.jetbrains.compose.resources.painterResource
 
 @Composable
-fun MovieItem(movie: MovieVO?,onTapMovie : (Int) -> Unit) {
+fun MovieItem(movie: MovieVO?,onTapMovie : (Long) -> Unit) {
     AsyncImage(
         movie?.getFullMovieBackdropPath(),
         contentDescription = null,
@@ -27,7 +27,7 @@ fun MovieItem(movie: MovieVO?,onTapMovie : (Int) -> Unit) {
             .size(width = MOVIE_ITEM_WIDTH, height = MOVIE_ITEM_HEIGHT)
             .clip(RoundedCornerShape(MARGIN_MEDIUM)).clickable(
                 onClick = {
-                    movie?.let { onTapMovie(it.id.toInt()) }
+                    movie?.let { onTapMovie(it.id) }
                 }
             )
